+++ ChangeLog.md
@@ -63,3 +63,4 @@
 
 6/10/2024
 - Added abandoning quests
+- Fixed quests not being complete when you accept them after getting the items